+++ packages/common/src/media/icons.ts
@@ -1,0 +1,2 @@
+export const seiLogoPNGUrl = 'https://raw.githubusercontent.com/cosmos/chain-registry/master/sei/images/sei.png';
+export const seiLogoSVGUrl = 'https://raw.githubusercontent.com/cosmos/chain-registry/master/sei/images/sei.svg';