+++ packages/cosmjs/src/cosmoskit/chains.ts
@@ -1,0 +1,129 @@
+export const SEI_BASE_COSMOS_KIT_CHAIN = {
+	bech32_config: {
+		bech32PrefixAccAddr: 'sei',
+		bech32PrefixAccPub: 'seipub',
+		bech32PrefixValAddr: 'seivaloper',
+		bech32PrefixValPub: 'seivaloperpub',
+		bech32PrefixConsAddr: 'seivalcons',
+		bech32PrefixConsPub: 'seivalconspub'
+	},
+	bech32_prefix: 'sei',
+	chain_name: 'sei',
+	daemon_name: 'seid',
+	description: 'Sei is the fastest Layer 1 blockchain, designed to scale with the industry.',
+	website: 'https://www.sei.io/',
+	fees: {
+		fee_tokens: [
+			{
+				denom: 'usei',
+				fixed_min_gas_price: 0.02,
+				low_gas_price: 0.02,
+				average_gas_price: 0.02,
+				high_gas_price: 0.04,
+				gas_costs: {
+					cosmos_send: 25000,
+					ibc_transfer: 50000
+				}
+			}
+		]
+	},
+	key_algos: ['secp256k1'],
+	keywords: ['sei'],
+	logo_URIs: {
+		png: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/sei/images/sei.png',
+		svg: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/sei/images/sei.svg'
+	},
+	network_type: 'mainnet',
+	node_home: '$HOME/.sei',
+	pretty_name: 'Sei',
+	slip44: 118,
+	status: 'live',
+	alternative_slip44s: [60],
+	staking: {
+		staking_tokens: [{ denom: 'usei' }]
+	},
+	images: [
+		{
+			png: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/sei/images/sei.png',
+			svg: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/sei/images/sei.svg'
+		}
+	]
+};
+
+export const PACIFIC_1_SEI_COSMOS_KIT_CHAIN = {
+	...SEI_BASE_COSMOS_KIT_CHAIN,
+	apis: {
+		rpc: [],
+		rest: [],
+		grpc: []
+	},
+	chain_id: 'pacific-1',
+	explorers: [
+		{
+			kind: 'seiscan',
+			url: 'https://www.seiscan.app/pacific-1',
+			tx_page: 'https://www.seiscan.app/pacific-1/txs/${txHash}',
+			account_page: 'https://www.seiscan.app/pacific-1/accounts/${accountAddress}'
+		},
+		{
+			kind: 'ping.pub',
+			url: 'https://ping.pub/sei',
+			tx_page: 'https://ping.pub/sei/tx/${txHash}',
+			account_page: 'https://ping.pub/sei/account/${accountAddress}'
+		},
+		{
+			kind: 'mintscan',
+			url: 'https://www.mintscan.io/sei',
+			tx_page: 'https://www.mintscan.io/sei/transactions/${txHash}',
+			account_page: 'https://www.mintscan.io/sei/accounts/${accountAddress}'
+		}
+	],
+	peers: {
+		seeds: [],
+		persistent_peers: []
+	}
+};
+
+export const ATLANTIC_2_SEI_COSMOS_KIT_CHAIN = {
+	...SEI_BASE_COSMOS_KIT_CHAIN,
+	apis: {
+		rpc: [],
+		rest: [],
+		grpc: []
+	},
+	chain_id: 'atlantic-2',
+	explorers: [
+		{
+			kind: 'seiscan',
+			url: 'https://www.seiscan.app/atlantic-2',
+			tx_page: 'https://www.seiscan.app/atlantic-2/txs/${txHash}',
+			account_page: 'https://www.seiscan.app/atlantic-2/accounts/${accountAddress}'
+		}
+	],
+	peers: {
+		seeds: [],
+		persistent_peers: []
+	}
+};
+
+export const ARCTIC_1_SEI_COSMOS_KIT_CHAIN = {
+	...SEI_BASE_COSMOS_KIT_CHAIN,
+	apis: {
+		rpc: [],
+		rest: [],
+		grpc: []
+	},
+	chain_id: 'arctic-1',
+	explorers: [
+		{
+			kind: 'seiscan',
+			url: 'https://www.seiscan.app/arctic-1',
+			tx_page: 'https://www.seiscan.app/arctic-1/txs/${txHash}',
+			account_page: 'https://www.seiscan.app/arctic-1/accounts/${accountAddress}'
+		}
+	],
+	peers: {
+		seeds: [],
+		persistent_peers: []
+	}
+};