+++ packages/cosmjs/src/index-browser.ts
@@ -1,4 +1,3 @@
-// eslint-disable @typescript-eslint/no-var-requires
 // Check if in browser environment
 if (typeof self !== 'undefined') {
 	const process = require('process');