+++ packages/cosmjs/src/index.ts
@@ -1,5 +1,5 @@
 export * from './core';
+export * from './cosmoskit';
 export * from './metamask-snap';
-export * from './react';
 export * from './utils';
 export * from './wallet';