+++ packages/cosmjs/src/react/components/WalletSelectModal/types.ts
@@ -1,9 +1,0 @@
-import { SeiWallet } from '../../../wallet';
-
-export type WalletSelectModalProps = {
-	wallets?: SeiWallet[];
-	classNameOverrides?: {
-		background?: string;
-		card?: string;
-	};
-};