+++ packages/cosmjs/src/react/hooks/useSigningCosmWasmClient/index.ts
@@ -1,2 +1,0 @@
-export { default as useSigningCosmWasmClient } from './useSigningCosmWasmClient';
-export type { UseSigningCosmWasmClient } from './useSigningCosmWasmClient';