+++ packages/cosmjs/src/react/hooks/useStargateClient/index.ts
@@ -1,2 +1,0 @@
-export { default as useStargateClient } from './useStargateClient';
-export type { UseStargateClient } from './useStargateClient';