+++ packages/cosmjs/src/utils/index.ts
@@ -1,4 +1,5 @@
-export { verifyArbitrary, makeADR36AminoSignDoc } from './signing';
-export { toBech32 } from './bech32';
-export { isValidSeiAddress, compressedPubKeyToAddress } from './address';
-export { serializeAminoSignDoc, serializeDirectSignDoc } from './serialize';
+export * from './address';
+export * from './bech32';
+export * from './hash';
+export * from './serialize';
+export * from './signing';