+++ packages/proto/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 0.0.0-internal-evm-20240301223227
+
+### Patch Changes
+
+- Removed react hooks and components, added exports for cosmoskit
+
+## 0.0.0-internal-evm-20240229151349
+
+### Patch Changes
+
+- Fixed missing typescript types, switched proto to use long instead of bigint
+
+## 0.0.0-internal-evm-20240229072551
+
+### Patch Changes
+
+- (WIP) Initial test of v2 packages
+
 ## 3.1.0
 
 ### Minor Changes